% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drink.R
\name{drink}
\alias{drink}
\title{Calculate how much more you need to drink}
\usage{
drink(shots)
}
\arguments{
\item{shots}{count of shot(s) already taken}
}
\description{
Calculate how much more you need to drink
}
\examples{
drinK(shots = 3)

}
